+++ src/espam/datamodel/graph/cnn/neurons/simple/NonLinear.java
@@ -259,7 +259,7 @@ public class NonLinear extends Neuron implements ConnectionDependent {
      * and DNN data formats are calculated
      */
     protected void setOperatorTimeComplexity(int inputChannels, int outputChannels){
-        int timeComplexity = 1;
+        long timeComplexity = 1;
         if(!(getInputDataFormat()==null)){
             timeComplexity = inputChannels * getInputHeight() * getInputWidth();
         }
